Parse units to s-expressions
This is a bit of a hack or rough sketch of how it could work, with some
deficiencies. Units are emitted in a different format than parsed, and
not all units can be emitted.

Originally we thought about parsing numbers with units as reader macros
expanding to ordinary s-expressions containing numbers and symbols, but
that may be more complicated than this "dumb" approach of defining a new
data type.