Sound A4 by default
This matches the default note sounded by the intern-apocalypse tool, but
it appears to be a mistake in that tool, which uses a constant named
'middleC' as the default note.

All sounds appear to render very close to how they originally rendered,
so this phase is complete, and we can move on to the next phase in the
project.

This is about 5.5kloc of Rust, which reimplements about 3.2kloc of
TypeScript. To render all of the audio programs in in intern-apocalypse,
this program runs (built in release mode) takes about 70-80ms, and the
original program takes about 2400ms. This is rougly a 32x speedup.